Prompt Generation Pipeline Improvements (#135)
* Release 2.0.3 (#134)

Bittensor 8.5.1

* enhancing prompts by adding conveyed motion with llama

* Mining docs fix setup_miner_env.sh -> setup_env.sh
