+++ bitmind/synthetic_data_generation/__init__.py
@@ -1,1 +1,2 @@
 from .synthetic_data_generator import SyntheticDataGenerator
+from .prompt_generator import PromptGenerator
