+++ docs/Mining.md
@@ -37,8 +37,8 @@ Install the remaining necessary requirements with the following chained command.
 ```bash
 conda activate bitmind
 export PIP_NO_CACHE_DIR=1
-chmod +x setup_miner_env.sh 
-./setup_miner_env.sh
+chmod +x setup_env.sh 
+./setup_env.sh
 ```
 
 ### Data
